I need your help to imagine how we could achieve this.
in the Dashboard, we want to help the user with some topics like :

- Do you have a sugar craving ? here is what to do.
- Are you looking for a late night snack ? here is what to do.

The target here is to help the user to avoid sugar cravings and late night snack by providing some insights on how to avoid, or limit or what other alternative is possible.

What i imagine is, a fancy display of the two topics listed above, that the user can click and will receive a pop up window with the content.

About the content :

- for sugar cravings, i want you to add informations from the internet on why we get sugar cravings and how to fight sugar cravings and which other alternatives are there.

- for late night snacks, smae things, from the internet why do crave food at night, how we fight and what alternatives are there.

for both pop ups, please keep the information precise and very short, with possiblity for the user to click on a button to extend and read more if he wishes