in the dashboard, card "daily habits checklist". I want to add some gamefication and start awarding rewards when goals are achieved.
Every time one of the checklist is checked, you get +1 reward. these rewards can then be used to gain prizes, like a cheat meal, a cheat day, a delicious dessert, an ice cream,..

What i am thinking about is the following :

- 15 rewards : a well deserved snack
- 35 rewards : dessert or ice cream
- 70 rewards : cheat meal
- 140 rewards : cheat day

if the user decides to use any of the prizes, the amount of rewards are deducted. For example, user has 42 rewards, decide to get Ice cream, rewards left are 7.

I give you the freedom to imagine and come up with a design on how to include this in the App