in the log food page, I want to re-write the code for log your food card and food item search.
I am not happy of how the search works, it is slow, complex and not precise.

I want to re-write it from scratch. which API do you suggest to use for food item search, and can you use this API to search for food ?

what is important is the following :

- the search result should contain the word typed.
- the exact word typed should be the first result showed and then show similar results with the word in it

all the other functionnalities remain as they are